backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL", "")
SUPABASE_KEY = os.getenv("SUPABASE_KEY", "")
OPENROUTER_API_KEY = os.getenv("OPENROUTER_API_KEY", "")
JWT_SECRET = os.getenv("JWT_SECRET", "changeme-secret")
ADMIN_PASSWORD = os.getenv("ADMIN_PASSWORD", "changeme")
