backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-multipart
pyjwt
httpx
python-dotenv
